requirements/test.txt
-r ./prod.txt
aioresponses==0.7.4
mock-5.1.0
pytest==7.0.0
pytest-asyncio==0.21.1
pytest-aiohttp==1.0.4